feat(TYPE): ensure ProductTypeController returns 404 if product type is not found on delete

+++ src/modules/product-type/controllers/product-type.controller.ts
@@ -62,6 +62,11 @@ export class ProductTypeController {
   @HttpCode(200)
   @HttpCode(404)
   async delete(@Param('id', ParseIntPipe) id: number) {
+    const productType = await this.service.readById(id);
+
+    if (!productType) {
+      throw new NotFoundException('Tipo de produto não encontrado');
+    }
     await this.service.delete(id);
     return 'Tipo de produto excluído com sucesso';
   }
